+++ apis/vara-eth/test/router.test.ts
@@ -42,15 +42,28 @@ afterAll(async () => {
 
 describe('router', () => {
   describe('upload code', () => {
-    test('should request code validation', async () => {
-      const tx = await router.requestCodeValidation(code);
-      codeId = tx.codeId;
-      const receipt = await tx.sendAndWaitForReceipt();
-      expect(receipt.blockHash).toBeDefined();
-      codeValidatedPromise = tx.waitForCodeGotValidated();
-    }, 60_000);
+    test(
+      'should request code validation',
+      async () => {
+        const tx = await router.requestCodeValidation(code);
+        codeId = tx.codeId;
+        const receipt = await tx.sendAndWaitForReceipt();
+        const transaction = await publicClient.getTransaction({ hash: receipt.transactionHash });
 
-    test.skip(
+        expect(transaction.blobVersionedHashes).toBeDefined();
+        expect(transaction.blobVersionedHashes!.length).toBeGreaterThan(0);
+
+        for (let i = 0; i < transaction.blobVersionedHashes!.length; i++) {
+          expect(transaction.blobVersionedHashes![i]).toBe(tx.blobVersionedHashes[i]);
+        }
+
+        expect(receipt.blockHash).toBeDefined();
+        codeValidatedPromise = tx.waitForCodeGotValidated();
+      },
+      config.longRunningTestTimeout * 3,
+    );
+
+    test(
       'should wait when code got validated',
       async () => {
         expect(await codeValidatedPromise).toBeTruthy();